+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-   Need to patch `ThingListGroupHelper.Includes` as well, which is involved in deciding what is a food source.
+
 ## [0.1.0] - 2024-04-30
 
 ### Added